# docs/swagger.yaml
basePath: /
definitions:
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Event:
    properties:
      age_limit:
        type: integer
      capacity:
        type: integer
      categories:
        description: relational fields
        items:
          type: string
        type: array
      created_at:
        type: string
      description:
        type: string
      dress_code:
        type: string
      end_time:
        type: string
      id:
        type: string
      image_url:
        description: non-mandatory fields
        type: string
      location:
        type: string
      name:
        type: string
      organizer:
        type: string
      participants:
        items:
          type: string
        type: array
      price:
        type: number
      start_time:
        type: string
      total_likes:
        type: integer
      updated_at:
        type: string
    type: object
  models.Level:
    enum:
    - country
    - state
    - city
    type: string
    x-enum-varnames:
    - Country
    - State
    - City
  models.Location:
    properties:
      created_at:
        type: string
      id:
        type: string
      level:
        $ref: '#/definitions/models.Level'
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      event_id:
        type: string
      id:
        type: string
      rating:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      joined_event:
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
  payload.CreateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  payload.CreateLocationRequest:
    properties:
      level:
        type: string
      name:
        type: string
    required:
    - level
    - name
    type: object
  payload.CreateResponse:
    properties:
      id:
        type: string
    type: object
  payload.GeneralResponse:
    properties:
      message:
        type: string
    type: object
  payload.GetCategoriesResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Category'
        type: array
    type: object
  payload.GetEventsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Event'
        type: array
    type: object
  payload.GetLocationsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Location'
        type: array
    type: object
  payload.GetReviewsResponse:
    properties:
      average_rating:
        type: number
      data:
        items:
          $ref: '#/definitions/models.Review'
        type: array
    type: object
  payload.LoginResponse:
    properties:
      token:
        type: string
    type: object
  payload.UpdateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  payload.UpdateLocationRequest:
    properties:
      level:
        type: string
      name:
        type: string
    type: object
  payload.UploadEventImageResponse:
    properties:
      url:
        type: string
    type: object
host: localhost:9999
info:
  contact: {}
  description: A matchmaking service API in Go using Gin framework.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Match-Event Backend API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Login
      parameters:
      - in: formData
        name: email
        required: true
        type: string
      - in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.LoginResponse'
      summary: Login
      tags:
      - users
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Register a new user
      parameters:
      - in: formData
        name: email
        required: true
        type: string
      - in: formData
        name: password
        required: true
        type: string
      - in: formData
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Register a new user
      tags:
      - users
  /categories:
    get:
      consumes:
      - application/json
      description: Get all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GetCategoriesResponse'
      summary: Get all categories
      tags:
      - categories
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Create Category
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payload.CreateResponse'
      summary: Create a new category
      tags:
      - categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Delete a category
      tags:
      - categories
    put:
      consumes:
      - application/json
      description: Update a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Category
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Update a category
      tags:
      - categories
  /events:
    get:
      consumes:
      - application/json
      description: Get all events
      parameters:
      - description: Event name
        in: query
        name: name
        type: string
      - description: Event categories
        example: cat1,cat2
        in: query
        name: categories
        type: string
      - description: Event locations
        example: loc_a,loc_b
        in: query
        name: locations
        type: string
      - description: Event price start
        in: query
        name: price_start
        type: string
      - description: Event price end
        in: query
        name: price_end
        type: string
      - description: Event age limit
        in: query
        name: age_limit
        type: string
      - description: Event start time
        example: "2023-12-02T20:00:00.000Z"
        in: query
        name: start_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GetEventsResponse'
      summary: Get all events
      tags:
      - events
    post:
      consumes:
      - multipart/form-data
      description: Create an event.
      parameters:
      - description: Image file
        in: formData
        name: image
        type: file
      - in: formData
        name: age_limit
        type: integer
      - in: formData
        name: capacity
        required: true
        type: integer
      - collectionFormat: csv
        description: 'Example: cat1,cat2'
        in: formData
        items:
          type: string
        name: categories
        type: array
      - in: formData
        name: description
        required: true
        type: string
      - in: formData
        name: dress_code
        type: string
      - description: 'Example: 2023-12-01T20:00:00.000Z'
        in: formData
        name: end_time
        required: true
        type: string
      - in: formData
        name: location
        required: true
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: organizer
        type: string
      - in: formData
        name: price
        required: true
        type: number
      - description: 'Example: 2023-12-01T20:00:00.000Z'
        in: formData
        name: start_time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.CreateResponse'
      summary: Create an event
      tags:
      - events
  /events/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Delete an event
      tags:
      - events
    get:
      consumes:
      - application/json
      description: Get an event by ID
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Event'
      summary: Get an event by ID
      tags:
      - events
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Update an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - in: formData
        name: age_limit
        required: true
        type: integer
      - in: formData
        name: capacity
        required: true
        type: integer
      - collectionFormat: csv
        description: 'Example: cat1,cat2'
        in: formData
        items:
          type: string
        name: categories
        required: true
        type: array
      - in: formData
        name: description
        required: true
        type: string
      - in: formData
        name: dress_code
        required: true
        type: string
      - description: 'Example: 2023-12-01T20:00:00.000Z'
        in: formData
        name: end_time
        required: true
        type: string
      - in: formData
        name: location
        required: true
        type: string
      - in: formData
        name: name
        required: true
        type: string
      - in: formData
        name: organizer
        required: true
        type: string
      - in: formData
        name: price
        required: true
        type: number
      - description: 'Example: 2023-12-01T20:00:00.000Z'
        in: formData
        name: start_time
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Update an event
      tags:
      - events
  /events/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.UploadEventImageResponse'
      summary: Upload an image
      tags:
      - events
  /events/{id}/join:
    post:
      consumes:
      - application/json
      description: Join an event
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Join an event
      tags:
      - events
  /events/popular:
    get:
      consumes:
      - application/json
      description: Get popular events
      parameters:
      - description: Top K
        in: query
        name: topK
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GetEventsResponse'
      summary: Get popular events
      tags:
      - events
  /locations:
    get:
      consumes:
      - application/json
      description: Get locations
      parameters:
      - description: Location level
        in: query
        name: level
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GetLocationsResponse'
      summary: Get locations
      tags:
      - locations
    post:
      consumes:
      - application/json
      description: Create a new location
      parameters:
      - description: Create Location
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.CreateLocationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payload.CreateResponse'
      summary: Create a new location
      tags:
      - locations
  /locations/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Delete a location
      tags:
      - locations
    put:
      consumes:
      - application/json
      description: Update a location
      parameters:
      - description: Location ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Location
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/payload.UpdateLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Update a location
      tags:
      - locations
  /reviews/{event_id}:
    get:
      consumes:
      - application/json
      description: Get reviews of an event
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GetReviewsResponse'
      summary: Get reviews of an event
      tags:
      - reviews
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a review
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Event ID
        in: path
        name: event_id
        required: true
        type: string
      - in: formData
        name: comment
        required: true
        type: string
      - in: formData
        name: rating
        required: true
        type: integer
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/payload.CreateResponse'
      summary: Create a review
      tags:
      - reviews
  /reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a review
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - description: Review ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Delete a review
      tags:
      - reviews
  /users:
    get:
      consumes:
      - application/json
      description: Get user info
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get user info
      tags:
      - users
    put:
      consumes:
      - application/x-www-form-urlencoded
      description: Update a user
      parameters:
      - description: With the bearer started
        in: header
        name: Authorization
        required: true
        type: string
      - in: formData
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payload.GeneralResponse'
      summary: Update a user
      tags:
      - users
  /users/{id}:
    get:
      consumes:
      - application/json
      description: Get user by id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get user by id
      tags:
      - users
swagger: "2.0"