# handlers/users/register.go
package users

import (
	"backend-go/models"
	payload "backend-go/payload/request"
	"backend-go/repository"
	"net/http"

	"github.com/gin-gonic/gin"
	"golang.org/x/crypto/bcrypt"
)

// Register godoc
// @Summary Register a new user
// @Description Register a new user
// @Tags users
// @Accept  x-www-form-urlencoded
// @Produce  json
// @Param body formData payload.RegisterRequest true "User"
// @Success 200 {object} payload.GeneralResponse
// @Router /auth/register [post]
func Register(c *gin.Context) {
	c.Header("Content-Type", "application/json")

	var body payload.RegisterRequest
	if c.ShouldBind(&body) != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"message": "Invalid request body",
		})
		return
	}

	doc, err := repository.GetUserByEmail(body.Email)
	if doc != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"message": "Email has been taken!",
		})
		return
	} else if err != nil && err.Error() != "User not found" {
		c.JSON(http.StatusInternalServerError, gin.H{"message": err.Error()})
		return
	}

	hash, err := bcrypt.GenerateFromPassword([]byte(body.Password), bcrypt.DefaultCost)
	if err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"message": "Failed to hash password",
		})
		return
	}

	user := models.User{
		Username: body.Username,
		Email:    body.Email,
		Password: string(hash),
	}
	_, err = repository.CreateUser(&user)

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"message": err.Error()})
		return
	}

	c.JSON(http.StatusOK, gin.H{
		"message": "User created successfully",
	})
}